+++ BasicBlock/testBasicBlock.py
@@ -15,7 +15,7 @@ import json
 import yaspc.Instruction.instruction as instruction
 import yaspc.BasicBlock.BasicBlock as BasicBlock
 import yaspc.BasicBlock.ConstructBasicBlock as ConstructBasicBlock
-import yaspc.IRParser.irParser as irParser
+import yaspc.IR_IO.irParser as irParser
 import yaspc.BasicBlock.ud as ud
 import yaspc.BasicBlock.PeepholeOptimization as PeepholeOptimization
 
@@ -24,7 +24,6 @@ def main():
     '''
     run a example of parser Three-address code and generate Basicblock list
     '''
-    inst_list = []
     with open('udtest.json', 'r') as input_file:
         ir_str = input_file.read()
         ir_json = json.loads(ir_str)
@@ -33,8 +32,7 @@ def main():
     # print(json.dumps(ir_json, sort_keys=True, indent=4))
     # print(json.dumps(ir_json['body'], sort_keys=True, indent=4))
     print(json.dumps(ir_json['functionlist'][0]['body'], sort_keys=True, indent=4))
-    for inst_dict in ir_json['functionlist'][0]['body']:
-        inst_list.append(irParser.parse_single_inst_from_json(inst_dict))
+    inst_list = irParser.decode_body(ir_json['functionlist'][0]['body'])
     print('function 0 body parse successfully.\n')
     for ith, inst in enumerate(inst_list):
         inst.lineth = ith
