+++ IR_IO/dsq.pas
